Renamed PaintDotNet to PaintDotNetLegacy. It's will be useful to see in what ways the legacy API is broken, if it's broken.
