+++ PaintDotNetLegacy/App.config
