+++ PaintDotNetLegacy/Properties/AssemblyInfo.cs
